+++ src/docstrings.jl
@@ -1212,13 +1212,14 @@ julia> @chain dt(db, df, "df_view") begin
 
 const docstring_copy_to =
 """
-       copy_to(conn, df_or_path, "name")
+       copy_to(conn, df_or_path, "name"; overwrite = false )
 Allows user to copy a df to the database connection. Currently supports DuckDB, SQLite, MySql
 
 # Arguments
 - `conn`: the database connection
 - `df`: dataframe to be copied or path to serve as source. With DuckDB, path supports .csv, .json, .parquet
 - `name`: name as string for the database to be used
+- `overwrite`: whether or not table with chosen name should be overwritten if it exists, defaults to `false`
 # Examples
 ```jldoctest
 julia> df = DataFrame(id = [string('A' + i ÷ 26, 'A' + i % 26) for i in 0:9],